Fix timestamp from FTP adapter
What does this Pull Request Do?
-------------------------------
The way the timestamp is calculated
```php
$timestamp = strtotime($month.' '.$day.' '.$time);
```

forgets that the FTP server don't show the year if the date is +-6 months. So, was calculating a wrong timestamp somethimes.

What I did was ask for the timestamp to the FTP server using
```php
$timestamp = ftp_mdtm($this->getConnection(), $name);
```
This solves the problem. Now you can do things like get the last file in a `listContents()` safely.

How should this be manually tested?
-----------------------------------
I added test to this.

Any background context?
-----------------------
I need for a project to search for the last file in many FTP accounts. This week I started using the library(which is awesome, thanks!!) and started to have issues about the files I was downloading and processing. Finally saw the issue, the `timestamp`, and here one way I propose to fix it.

Any extra info?
---------------
This fix has a performance issue if the directory have lots of files, becouse, instead of using only one LIST, we'll be adding one MODTIME call to the server for each file. So, we'll end with `N` roundtrips more (`N` number of files in the directory).

Options:
Another option would be to check the date and see if it's +-6 month, and fix it if necessary. It's easy if we get the current date of the server we are running, but to do it properly we should use the date of the FTP server we are in. We can do this by writing a file to check the FTP server's date, but (1) we could be breaking read-only implementa and (2) if we do it do it in the `normalizeObject()` method is as bad for performance (3) if we do it before normalizing, we have to carry the date with us, changing the parameters we actually have for all the other methods until we get to `normalizeObject()`.

Hope you find this useful! ;)