allow prefixes like URI schemes in MountManager
Allow prefixes that are valid schemes according to RFC3986. That is, a
prefix must start with a letter followed by any combination of letters,
digits, plus (`+`), period (`.`), or hyphen (`-`). This change allows
case insensitive usage of prefixes as the internal key for the filesystems
array is lowercased.

See http://tools.ietf.org/html/rfc3986#section-3.1

This change is a breaking change as prior to this the mounting of
filesystems never checked for valid prefixes and prior to this change
a all number prefix (`1` or `1337`) was valid.