Avoid PHP Warning on FTP Connection
If `$this->connection` has failed and is a boolean, calling `ftp_systype($this->connection)` throws a PHP warning (
Warning: ftp_systype() expects parameter 1 to be resource, boolean given). I propose first checking to see if $this->connection is a resource or a boolean/null and if it's going to fail anyway, to start the failure sooner.