Improve isConnected()
`isConnected()` always lists the root directory of the ftp server to determine if the connection is active. 

This changes that command to list the root of the adapter for that check instead. It's useful for when you have a very large root directory causing LIST commands to take a long time and you want to specify a sub-directory as root to speed things up.