FTP: Update resource to allow new FTPConnection objects
In PHP 8.1, `ftp` resources are [migrated to `FTPConnection` class objects](https://php.watch/versions/8.1/FTPConnection-resource). This PR updates all `is_resource` checks and DocBlock types to accept/indicate the new class objects.

This is probably a PR to send quite early, but given the extreme popularity of flysystem, I think there will be people running their builds in PHP nightly builds. This change in PHP was done a few weeks ago, and is part of the resource of object migration plan, hence there was no RFCs or lengthy discussions.