Avoid type errors when public key is not retrieved
When dealing with unstable connections, phpseclib may return `false`
when retrieving the public key.  That causes `null` to be passed down to
`base64_decode()` and triggers a TypeError due to strict types.

This exits early, preventing triggering the problem. However, testing is
a bit tricky since we need an unstable connection.